Split fingerprint file into smaller for better overview
