+++ scikit_mol/fingerprints/__init__.py
@@ -1,0 +1,15 @@
+from .baseclasses import (
+    FpsTransformer,
+    FpsGeneratorTransformer,
+)  # TODO, for backwards compatibility with tests, needs to be removed
+
+from .atompair import AtomPairFingerprintTransformer, AtomPairFPGeneratorTransformer
+from .avalon import AvalonFingerprintTransformer
+from .maccs import MACCSKeysFingerprintTransformer
+from .minhash import MHFingerprintTransformer, SECFingerprintTransformer
+from .morgan import MorganFingerprintTransformer, MorganFPGeneratorTransformer
+from .rdkitfp import RDKitFingerprintTransformer, RDKitFPGeneratorTransformer
+from .topologicaltorsion import (
+    TopologicalTorsionFingerprintTransformer,
+    TopologicalTorsionFPGeneatorTransformer,
+)
